Create UpdatedFlow
This updated flow is much more reliable than the last and much simpler in every aspect

This Flow writes and reads to the timers file in home/pi/ so you may need to create it first (touch timers)
because it physically writes a file, then reads it every minute the set timers survives reboots. 

The reading of the timers file every minute maybe why this flow is more reliable than the last.
I couldn't identify the old flows glitch so this is the much improved refresh.